+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.6.7
 ## mm/dd/2019
 
+1. [](#new)
+    * Added a new `bin/grav yamllinter` CLI command to find YAML Linting issues [#2468](https://github.com/getgrav/grav/issues/2468#issuecomment-485151681)
 1. [](#improved)
     * Improve `FormTrait` backwards compatibility with existing forms
     * Added a new `Utils::getSubnet()` function for IPv6 parsing [#2465](https://github.com/getgrav/grav/pull/2465)